+++ CHANGELOG.md
@@ -8,7 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 
-- `onStream()` and `createStream()` support for streaming data between `Messenger` instances on separate threads.
+- `onStream()` and `createStream()` support for streaming data between `Messenger` instances on separate threads + docs.
 
 ### Changed