+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## [0.2.4] - 2022-08-12
+
 ### Added
 
 - `onStream()` and `createStream()` support for streaming data between `Messenger` instances on separate threads + docs.